RENAME #17 Game and Game.config
